Move Cube mesh data into real GPU buffers, read via BDA pointers
Index/vertex/normal data no longer lives in static shader arrays — it's
allocated with GpuMalloc(..., Upload) once in OnInit and read in Cube.slang
through genuine buffer-device-address pointer fields in the push constants
(Vertex* vertices; uint32_t* indices;), dereferenced manually by
SV_VertexID. No Vulkan-level index/vertex buffer binding; still a plain
Draw(cmd, 36). Expanded from 8 shared corners to 24 unique vertices (4 per
face) since flat per-face normals can't live on a shared corner. Enabled
scalarBlockLayout (core since Vulkan 1.2) for the tightly-packed
{float3 position; float3 normal;} pointee layout Slang emits by default.

Also fixes a real bug found while wiring this up: GpuMalloc<T>(data, usage)
and friends called their own plain u32-size overload with sizeof(T), a
size_t — an exact match for the template itself (deducing T=size_t), which
beats the plain overload's narrowing u32 conversion in overload resolution.
That's not a corner case; it's what "GpuMalloc(sizeof(x), usage)" always
does, and it recursed into itself until the stack overflowed. Fixed with
explicit u32 casts at the recursive call sites and at CubeApp's own use of
GpuMalloc for the new mesh buffers.

Confirmed live: no crash, zero new validation errors, push-constant byte
counts match the new struct exactly, and the rendered cube is pixel-
equivalent to the prior working baseline.

+++ PhxEngine/RHI/RHI.h
@@ -90,23 +90,33 @@ namespace phx::rhi
     // Valid only for the frame it was allocated in; never freed individually.
     [[nodiscard]] GpuAllocation GpuTempMalloc(u32 size);
 
+    // Note: every sizeof(T) below is explicitly cast to u32. sizeof() is
+    // size_t (8 bytes); calling GpuMalloc/GpuTempMalloc with a bare size_t
+    // is an *exact* match for these very templates (deducing T=size_t) —
+    // beating the plain u32-size overloads, which need a narrowing
+    // conversion and so lose the overload-resolution tiebreak. Without the
+    // cast, that recurses into itself infinitely instead of calling the
+    // intended plain allocator. The explicit u32 makes it an exact-match
+    // tie instead, which the non-template overload wins by the standard
+    // "prefer non-template on a tie" rule.
+
     template<typename T>
     [[nodiscard]] GpuAllocation GpuMalloc()
     {
-        return GpuMalloc(sizeof(T));
+        return GpuMalloc(static_cast<u32>(sizeof(T)));
     }
 
     template<typename T>
     [[nodiscard]] GpuAllocation GpuTempMalloc()
     {
-        return GpuTempMalloc(sizeof(T));
+        return GpuTempMalloc(static_cast<u32>(sizeof(T)));
     }
 
     // Allocates and writes `data` in one call.
     template<typename T>
     [[nodiscard]] GpuAllocation GpuTempMalloc(const T& data)
     {
-        GpuAllocation alloc = GpuTempMalloc(sizeof(T));
+        GpuAllocation alloc = GpuTempMalloc(static_cast<u32>(sizeof(T)));
         if (alloc.cpu_ptr)
             std::memcpy(alloc.cpu_ptr, &data, sizeof(T));
         return alloc;
@@ -118,7 +128,7 @@ namespace phx::rhi
     template<typename T>
     [[nodiscard]] GpuAllocation GpuMalloc(const T& data, GpuMemoryUsage usage = GpuMemoryUsage::Upload)
     {
-        GpuAllocation alloc = GpuMalloc(sizeof(T), usage);
+        GpuAllocation alloc = GpuMalloc(static_cast<u32>(sizeof(T)), usage);
         PHX_ASSERT(alloc.cpu_ptr && "GpuMalloc<T> with a value needs a host-visible usage (Upload/ReadBack) — DeviceLocal has no cpu_ptr to write through.");
         if (alloc.cpu_ptr)
             std::memcpy(alloc.cpu_ptr, &data, sizeof(T));
